requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
pillow>=9.5.0